Added more functionality. like allowing singned in members to edit their info Enhanced user experience by disabling some potions of the UI when a user wasn't logged in Added timestamps to database tables Added loan clearance functionality Added member active/ inactive ability for accounts Still there are lingering errors to fix, but ill fix em soon. like the one for detecting loan overpayments

+++ src/com/sacco/classes/Admin.java
@@ -9,6 +9,7 @@ import static com.sacco.classes.Member.getId;
 import static com.sacco.classes.Member.isAdmin;
 import java.sql.SQLException;
 import java.sql.Statement;
+import javax.security.auth.login.AccountException;
 import javax.swing.JTextArea;
 
 /**
@@ -17,9 +18,14 @@ import javax.swing.JTextArea;
  */
 public class Admin extends Member {
 
-//    public Admin() {
-//        super();
-//    }
+    // handy, since it will be always checked on instance creation
+    public Admin() throws AccountException {
+        // check if the user is who they claim to be
+        if (!Member.isAdmin()) {
+            throw new AccountException("Are you really an admin?. Check again");
+        }
+    }
+
     // check login status. for administrative use since any id can be passed,
     // and a basic user shouldn't be able to see other user's status
     protected static boolean CheckIfLoggedIn(long id) {
@@ -36,7 +42,7 @@ public class Admin extends Member {
     }
 
     // an Admin task. of course
-    public boolean DeleteMember(long id) throws SQLException {
+    public boolean DeleteMember(long id) throws SQLException, AccountException {
         if (isAdmin()) {
             try {
                 //System.out.println(conn);
@@ -58,39 +64,61 @@ public class Admin extends Member {
                 close();
             }
         }
-        return false;
+        throw new AccountException("You are not allowed to perform this action");
     }
 
-    public void DisplayAllMembers(JTextArea jt) throws SQLException {
-        if (isAdmin()) {
-            jt.setText("");
-            try {
-                String sql = "SELECT * FROM sacco.members";
-                stmt = conn.prepareStatement(sql);
-                result = stmt.executeQuery();
-
-                jt.append("Here are all the members in the sacco \n\n");
-                jt.append("FIRSTNAME\t LASTNAME \t GENDER \t DATE_OF_BIRTH \t ADDRESS \t EMAIL\n\n");
-                while (result.next()) {
-                    // store logged in userID in mem
-                    setId(id);
-                    jt.append(result.getString("firstname"));
-                    jt.append("\t");
-                    jt.append(result.getString("lastname"));
-                    jt.append("\t");
-                    jt.append(result.getString("gender"));
-                    jt.append("\t");
-                    jt.append(result.getDate("dob").toString());
-                    jt.append("\t\t");
-                    jt.append(result.getString("address"));
-                    jt.append("\t");
-                    jt.append(result.getString("email"));
-                    jt.append("\n");
-                }
-            } finally {
-                close();
+    // the admin should be able to close another user's account
+    public boolean CloseAccount(long id) throws SQLException {
+        try {
+            //System.out.println(conn);
+            String sql = "UPDATE sacco.members SET active = ? WHERE id = ?";
+            stmt = conn.prepareStatement(sql);
+            stmt.setShort(1, Member.INACTIVE);
+            stmt.setLong(2, id);
+            int rows = stmt.executeUpdate();
+            if (rows == 0) {
+                // a user wasn't removed
+                throw new SQLException("Operation failed");
+            } else {
+                return true;
             }
+
+        } finally {
+            // close resources
+            close();
         }
+    }
+
+    public void DisplayAllMembers(JTextArea jt) throws SQLException, AccountException {
+
+        jt.setText("");
+        try {
+            String sql = "SELECT * FROM sacco.members";
+            stmt = conn.prepareStatement(sql);
+            result = stmt.executeQuery();
+
+            jt.append("Here are all the members in the sacco \n\n");
+            jt.append("FIRSTNAME\t LASTNAME \t GENDER \t DATE_OF_BIRTH \t ADDRESS \t EMAIL\n\n");
+            while (result.next()) {
+                // store logged in userID in mem
+                setId(id);
+                jt.append(result.getString("firstname"));
+                jt.append("\t");
+                jt.append(result.getString("lastname"));
+                jt.append("\t");
+                jt.append(result.getString("gender"));
+                jt.append("\t");
+                jt.append(result.getDate("dob").toString());
+                jt.append("\t\t");
+                jt.append(result.getString("address"));
+                jt.append("\t");
+                jt.append(result.getString("email"));
+                jt.append("\n");
+            }
+        } finally {
+            close();
+        }
+
     }
 
     public long makeMemberAdmin(long id) throws SQLException {
@@ -123,20 +151,24 @@ public class Admin extends Member {
 
     }
 
-    protected boolean DestroyMemberLoan(long id) throws SQLException {
-        try {
-            //System.out.println(conn);
-            String sql = "DELETE FROM sacco.Loans WHERE member_id = ?";
-            stmt = conn.prepareStatement(sql);
-            stmt.setLong(1, id);
+    public boolean ClearMemberLoan(long id) throws SQLException, AccountException {
+        if (isAdmin()) {
+            try {
+                //System.out.println(conn);
+                String sql = "DELETE FROM sacco.Loans WHERE member_id = ?";
+                stmt = conn.prepareStatement(sql);
+                stmt.setLong(1, id);
 
-            // PreparedStatement.setString(8, getPassword());
-            int rows = stmt.executeUpdate();
-            return rows == 0;
+                // PreparedStatement.setString(8, getPassword());
+                int rows = stmt.executeUpdate();
+                return rows == 0;
 
-        } finally {
-            // close resources
-            close();
+            } finally {
+                // close resources
+                close();
+            }
+        } else {
+            throw new AccountException("You are not allowed to perform this action");
         }
     }
